==> jupyter-notebooks/01_obtener_ids_followers_spotify.ipynb <==
# Importamos las librerías que necesitamos

# Librerías de extracción de datos
# -----------------------------------------------------------------------
from bs4 import BeautifulSoup
import requests
from tqdm import tqdm
# Tratamiento de datos
# -----------------------------------------------------------------------
import pandas as pd
import numpy as np
import datetime

# Importamos las librerías que necesitamos

# Librerías de extracción de datos
# -----------------------------------------------------------------------

# Importaciones:
# Beautifulsoup
from bs4 import BeautifulSoup

# Requests
import requests

import pandas as pd
import numpy as np

from time import sleep
import random
# Importar librerías para automatización de navegadores web con Selenium
# -----------------------------------------------------------------------
from selenium import webdriver  # Selenium es una herramienta para automatizar la interacción con navegadores web.
from webdriver_manager.chrome import ChromeDriverManager  # ChromeDriverManager gestiona la instalación del controlador de Chrome.
from selenium.webdriver.common.keys import Keys  # Keys es útil para simular eventos de teclado en Selenium.
from selenium.webdriver.support.ui import Select  # Select se utiliza para interactuar con elementos <select> en páginas web.
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import NoSuchElementException # Excepciones comunes de selenium que nos podemos encontrar 

import sys
sys.path.append("../")
import src.soporte_spotify as api

sp = api.load_credentials()
sp.current_user()

# Obtener datos de marcas
- Vamos a obtener los 1000 primeros usuarios de cada marca y extraeremos los géneros musicales que escuchan
    - Zara
    - Primark
    - Nike
    - Adidas
    - H&M
    - Pull & Bear
    - bershka
    - Stradivarius
    - Mango
    - Desigual
### Nota Importante
- Spotify no me deja acceder a los ids de usuarios que siguen a un usuario, pero investigando, en el html, ponen el id de usuario en la parte de "followers"
- No puedo ver por playlists los usuarios
- Pero puedo ver por marca el perfil de música que se escucha
- accediendo aqui con selenium o beautiful soup
    - https://open.spotify.com/user/r6ivwuv0ebk346hhxo446pbfv/followers (perfil zara)
    - No necesito iniciar sesión
- En el HTML puedo extraer los ids de usuarios

# Zara
- https://open.spotify.com/user/r6ivwuv0ebk346hhxo446pbfv/followers
- r6ivwuv0ebk346hhxo446pbfv

zara = api.obtener_html_followers("r6ivwuv0ebk346hhxo446pbfv")
df_zara = api.extraer_ids_usuario(sp,"Zara",zara)
df_zara.to_csv("../datos/01 Spotify/01_followers_zara.csv")
df_zara.head()

# Primark
- https://open.spotify.com/user/1u0ewgq4b77l1ttu6j4oyaylw
- user: 1u0ewgq4b77l1ttu6j4oyaylw

primark = api.obtener_html_followers("1u0ewgq4b77l1ttu6j4oyaylw")
df_primark = api.extraer_ids_usuario(sp,"Primark",primark)
df_primark.to_csv("../datos/01 Spotify/02_followers_primark.csv")
df_primark.head()

# Nike
- https://open.spotify.com/user/nikerunclub 
- user: nikerunclub 

nike = api.obtener_html_followers("nikerunclub")
df_nike = api.extraer_ids_usuario(sp,"Nike",nike)
df_nike.to_csv("../datos/01 Spotify/03_followers_nike.csv")
df_nike.head()

# Adidas
- https://open.spotify.com/user/430wexvtgeypb6zdn45ge9c76 
- 430wexvtgeypb6zdn45ge9c76 

adidas = api.obtener_html_followers("430wexvtgeypb6zdn45ge9c76")
df_adidas = api.extraer_ids_usuario(sp,"Adidas",adidas)
df_adidas.to_csv("../datos/01 Spotify/04_followers_adidas.csv")
df_adidas.head()

# H&M
- https://open.spotify.com/user/hm
- hm

hm = api.obtener_html_followers("hm")
df_hm = api.extraer_ids_usuario(sp,"H&M",hm)
df_hm.to_csv("../datos/01 Spotify/05_followers_hm.csv")
df_hm.head()

# Pull & Bear
- https://open.spotify.com/user/pull_and_bear
- pull_and_bear

pullbear = api.obtener_html_followers("pull_and_bear")
df_pullbear = api.extraer_ids_usuario(sp,"Pull & Bear",pullbear)
df_pullbear.to_csv("../datos/01 Spotify/06_followers_pull_and_bear.csv")
df_pullbear.head()

# bershka
- https://open.spotify.com/user/bershka
- bershka

bershka = api.obtener_html_followers("bershka")
df_bershka = api.extraer_ids_usuario(sp,"bershka",bershka)
df_bershka.to_csv("../datos/01 Spotify/07_followers_bershka.csv")
df_bershka.head()

# Stradivarius
- https://open.spotify.com/user/stradivas
- stradivas

stradivarius = api.obtener_html_followers("stradivas")
df_stradivarius = api.extraer_ids_usuario(sp,"Stradivarius",stradivarius)
df_stradivarius.to_csv("../datos/01 Spotify/08_followers_stradivarius.csv")
df_stradivarius.head()

# Mango
- https://open.spotify.com/user/mango_official
- mango_official

mango = api.obtener_html_followers("mango_official")
df_mango = api.extraer_ids_usuario(sp,"Mango",mango)
df_mango.to_csv("../datos/01 Spotify/09_followers_mango.csv")
df_mango.head()

# Desigual
- https://open.spotify.com/user/1119086036
- 1119086036

desigual = api.obtener_html_followers("1119086036")
df_desigual = api.extraer_ids_usuario(sp,"Desigual",desigual)
df_desigual.to_csv("../datos/01 Spotify/10_followers_desigual.csv")
df_desigual.head()
