==> jupyter-notebooks/02_obtener_listas_spotify.ipynb <==
# Tratamiento de datos
# -----------------------------------------------------------------------
import pandas as pd
import numpy as np
# Requests
import requests
from time import sleep
from tqdm import tqdm


import sys
sys.path.append("../")
import src.soporte_spotify as api

sp = api.load_credentials()
sp.current_user()

# Verificar si nos han bloqueado
- Tenemos un limite diario y si no spotify devuelve 429, algo que con spotipy no se ve, asi que usamos requests

token = api.request_token()
# 4. Endpoint para obtener playlists de un usuario
user_id = "lpet16qfnsa9zn97li98q8c50"  # Cambia por el ID del usuario de Spotify
url = f"https://api.spotify.com/v1/users/{user_id}/playlists"

headers = {
    "Authorization": f"Bearer {token}"
}

response = requests.get(url, headers=headers)


print(f"{response.status_code}, Detalles: {response.text}")
print(response.headers)


# Obtener Géneros que escuchan
- Vamos a obtener los generos que escuchan los followers que hemos almacenado

zara = pd.read_csv('../datos/01 Spotify/01_followers_zara.csv',index_col=0)
primark = pd.read_csv("../datos/01 Spotify/02_followers_primark.csv",index_col=0)
nike = pd.read_csv("../datos/01 Spotify/03_followers_nike.csv",index_col=0)
adidas = pd.read_csv("../datos/01 Spotify/04_followers_adidas.csv",index_col=0)
hm = pd.read_csv("../datos/01 Spotify/05_followers_hm.csv",index_col=0)
pullbear = pd.read_csv("../datos/01 Spotify/06_followers_pull_and_bear.csv",index_col=0)
bershka = pd.read_csv("../datos/01 Spotify/07_followers_bershka.csv",index_col=0)
stradivarius = pd.read_csv("../datos/01 Spotify/08_followers_stradivarius.csv",index_col=0)
mango = pd.read_csv("../datos/01 Spotify/09_followers_mango.csv",index_col=0)
desigual = pd.read_csv("../datos/01 Spotify/10_followers_desigual.csv",index_col=0)

# Zara

lista_playlists = []
for id in tqdm(zara["user_id"]):
    playlists = api.obtener_playlists(sp,id)
    lista_playlists.append(playlists)

zara["playlists"] = lista_playlists

### Eliminar usuarios sin playlists
- No nos aportan información 

zara = zara.loc[zara["playlists"] != "No playlists"].reset_index(drop=True)
zara.to_csv("../datos/01 Spotify/01_followers_zara.csv",index=True)

# Primark

lista_playlists = []
for id in tqdm(primark["user_id"]):
    playlists = api.obtener_playlists(sp,id)
    lista_playlists.append(playlists)
    sleep(0.1)

primark["playlists"] = lista_playlists

### Eliminar usuarios sin playlists
- No nos aportan información 

primark = primark.loc[primark["playlists"] != "No playlists"].reset_index(drop=True)
primark.to_csv("../datos/01 Spotify/02_followers_primark.csv",index=True)

primark.head()

# Nike

lista_playlists = []
for id in tqdm(nike["user_id"]):
    playlists = api.obtener_playlists(sp,id)
    lista_playlists.append(playlists)
    sleep(0.1)

nike["playlists"] = lista_playlists

### Eliminar usuarios sin playlists
- No nos aportan información 

nike = nike.loc[nike["playlists"] != "No playlists"].reset_index(drop=True)
nike.to_csv("../datos/01 Spotify/03_followers_nike.csv",index=True)

nike.head()

# Adidas

lista_playlists = []
for id in tqdm(adidas["user_id"]):
    playlists = api.obtener_playlists(sp,id)
    lista_playlists.append(playlists)
    sleep(0.1)

adidas["playlists"] = lista_playlists

### Eliminar usuarios sin playlists
- No nos aportan información 

adidas = adidas.loc[adidas["playlists"] != "No playlists"].reset_index(drop=True)
adidas.to_csv("../datos/01 Spotify/04_followers_adidas.csv",index=True)
adidas.head()

# H&M

lista_playlists = []
for id in tqdm(hm["user_id"]):
    playlists = api.obtener_playlists(sp,id)
    lista_playlists.append(playlists)
    sleep(0.1)
hm["playlists"] = lista_playlists

### Eliminar usuarios sin playlists
- No nos aportan información 

hm = hm.loc[hm["playlists"] != "No playlists"].reset_index(drop=True)
hm.to_csv("../datos/01 Spotify/05_followers_hm.csv",index=True)

# Pull & Bear

lista_playlists = []
for id in tqdm(pullbear["user_id"]):
    playlists = api.obtener_playlists(sp,id)
    lista_playlists.append(playlists)
    sleep(0.1)

pullbear["playlists"] = lista_playlists

### Eliminar usuarios sin playlists
- No nos aportan información 

pullbear = pullbear.loc[pullbear["playlists"] != "No playlists"].reset_index(drop=True)
pullbear.to_csv("../datos/01 Spotify/06_followers_pull_and_bear.csv",index=True)

# Bershka

lista_playlists = []
for id in tqdm(bershka["user_id"]):
    playlists = api.obtener_playlists(sp,id)
    lista_playlists.append(playlists)
    sleep(0.1)

bershka["playlists"] = lista_playlists

### Eliminar usuarios sin playlists
- No nos aportan información 

bershka = bershka.loc[bershka["playlists"] != "No playlists"].reset_index(drop=True)
bershka.to_csv("../datos/01 Spotify/07_followers_bershka.csv",index=True)

# Stradivarius

lista_playlists = []
for id in tqdm(stradivarius["user_id"]):
    playlists = api.obtener_playlists(sp,id)
    lista_playlists.append(playlists)

stradivarius["playlists"] = lista_playlists

### Eliminar usuarios sin playlists
- No nos aportan información 

stradivarius = stradivarius.loc[stradivarius["playlists"] != "No playlists"].reset_index(drop=True)
stradivarius.to_csv("../datos/01 Spotify/08_followers_stradivarius.csv",index=True)

# Mango

lista_playlists = []
for id in tqdm(mango["user_id"]):
    playlists = api.obtener_playlists(sp,id)
    lista_playlists.append(playlists)

mango["playlists"] = lista_playlists

### Eliminar usuarios sin playlists
- No nos aportan información 

mango = mango.loc[mango["playlists"] != "No playlists"].reset_index(drop=True)
mango.to_csv("../datos/01 Spotify/09_followers_mango.csv",index=True)

# Desigual
- Solo sacamos un máximo de 10 por usuario

lista_playlists = []
for id in tqdm(desigual["user_id"]):
    playlists = api.obtener_playlists(sp,id)
    lista_playlists.append(playlists)

desigual["playlists"] = lista_playlists

### Eliminar usuarios sin playlists
- No nos aportan información 

desigual = desigual.loc[desigual["playlists"] != "No playlists"].reset_index(drop=True)
desigual.to_csv("../datos/01 Spotify/10_followers_desigual.csv",index=True)
