==> jupyter-notebooks/05_obtener_generos_artistas.ipynb <==
# Tratamiento de datos
# -----------------------------------------------------------------------
import pandas as pd
import numpy as np
# Requests
import requests
from time import sleep
from tqdm import tqdm

# Para convertir de string a diccionario
import ast
import os
import json

import sys
sys.path.append("../")
import src.soporte_spotify as api

sp = api.load_credentials()
sp.current_user()

zara = pd.read_csv('../datos/01 Spotify/01 Resumen Marcas/resumen_zara.csv')
zara["unique_artists"] = zara["unique_artists"].apply(ast.literal_eval)
zara["artist_ranking"] = zara["artist_ranking"].apply(lambda x: ast.literal_eval(x) if pd.notna(x) else [])
zara

# Solicitar token y obtener artistas
token = api.request_token(silent=True)
generos = api.obtener_generos(token,zara["unique_artists"])

zara["genres"] = str(list(generos.keys()))
zara["genres_ranking"] = str(sorted(generos.items(), key=lambda x: x[1], reverse=True))
zara.to_csv("../datos/01 Spotify/01 Resumen Marcas/resumen_zara.csv",index=False)
