==> jupyter-notebooks/06_base_de_datos.ipynb <==
# Vamos a utilizar SupaBase
- Una Base de datos en la nube que usa PostgreSQL
- Allí almacenaremos los datos recopilados
- La idea es poder tener la app funcionando desde cualquier lugar, que sea pública
- pip install python-dotenv psycopg2

Todo lo de aquí lo pasaremos a Funciones para facilitar subir datos en caso de ser necesario

import sys
sys.path.append("../")
import src.soporte_sql as sql
import pandas as pd
import ast

conexion = sql.conectar_bd()

# Crear tablas
- Vamos a organizarlo por tablas

df = pd.read_csv("../datos/01 Spotify/01_followers_zara.csv")
df.sample()

df2 = pd.read_csv("../datos/01 Spotify/01 Resumen Marcas/resumen_zara.csv")
df2.sample()

# Tabla Brands
- Aquí almacenaremos los nombres de las marcas y su user id
### Crear Dataframe con formato tabla

dictio_marcas = {
    "zara" : "r6ivwuv0ebk346hhxo446pbfv",
    "primark" : "1u0ewgq4b77l1ttu6j4oyaylw",
    "nike" : "nikerunclub",
    "adidas" : "430wexvtgeypb6zdn45ge9c76",
    "h & m" : "hm",
    "pull & bear" : "pull_and_bear",
    "bershka" : "bershka",
    "stradivarius" : "stradivas",
    "mango" : "mango_official",
    "desigual" : "1119086036"
}
brands = pd.DataFrame({
    "brand" : list(dictio_marcas.keys()),
    "user_id" : list(dictio_marcas.values())
})

url_list = []
for id in brands["user_id"]:
    url_list.append(f"https://open.spotify.com/user/{id}")
brands["url"] = url_list
brands

### Crear tabla y subir los contenidos del DF

conexion = sql.conectar_bd()
query = '''CREATE TABLE brands (
            id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
            name VARCHAR UNIQUE NOT NULL,
            user_id VARCHAR UNIQUE NOT NULL,
            url VARCHAR UNIQUE NOT NULL
            );'''
sql.modificar_bd(conexion,query)

conexion = sql.conectar_bd()
query = '''INSERT INTO brands(name,user_id,url) VALUES (%s,%s,%s)'''
sql.insertar_muchos_datos(conexion,query,sql.generar_tupla(brands))

brands.to_csv("../datos/02 Base de Datos/tabla_brands.csv",index=False)

# Tabla Followers
- Aquí estará toda la información de los seguidores, y de que marca provienen

zara = pd.read_csv('../datos/01 Spotify/01_followers_zara.csv')
zara = zara[["brand","username","user_id"]]

### Preparar tabla
- Pasamos "brand" al id que tenemos en la tabla "brands" para llamara a la clave foránea

dictio_brand = {
    "Zara": 1,
    "Primark": 2,
    "Nike": 3,
    "Adidas": 4,
    "H&M": 5,
    "Pull & Bear": 6,
    "bershka": 7,
    "Stradivarius": 8,
    "Mango": 9,
    "Desigual": 10
}
zara["brand"] = zara["brand"].map(dictio_brand)
zara = zara[["username","user_id","brand"]]
zara.sample()

### Crear tabla y subir los contenidos del DF

conexion = sql.conectar_bd()
query = ''' CREATE TABLE followers (
            id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
            username VARCHAR NOT NULL,
            user_id VARCHAR NOT NULL,
            brand_id BIGINT REFERENCES brands(id) ON DELETE CASCADE
            );'''
sql.modificar_bd(conexion,query)

conexion = sql.conectar_bd()
query = '''INSERT INTO followers(username,user_id,brand_id) VALUES (%s,%s,%s)'''
sql.insertar_muchos_datos(conexion,query,sql.generar_tupla(zara))

zara.to_csv("../datos/02 Base de Datos/02_tabla_followers.csv",index=False)

# Tabla Playlists
- Esta tabla tendra cada playlist con su id, su nombre y la marca de donde procede
- además estarán referenciadas por brand y user_id

zara = pd.read_csv('../datos/01 Spotify/01_followers_zara.csv')

### Obtenemos el id de los usuarios haciendo una consulta a la bd

conexion = sql.conectar_bd()
query = '''SELECT * FROM followers'''
df_users = sql.consulta_sql(conexion,query)
df_users = df_users[["id","user_id","username"]]
df_users.sample()

### Realizamos merge y convertimos a diccionario "playlists"

unir = pd.merge(left=df_users,right=zara,on="user_id")
playlists = unir[["id","playlists"]]
playlists["playlists"] = playlists["playlists"].apply(ast.literal_eval)
playlists.sample()

### Creamos un Dataframe que:
- tendra el id del usuario a quien pertenece
- nombre de la playlist
- id de la playlist
- Uno por fila

id_list = []
playlist_names = []
playlist_ids = []

for index, fila in playlists.iterrows():
    user_id = fila["id"]
    playlist = fila["playlists"]

    for playlist_name, playlist_id in playlist.items():
        id_list.append(user_id)
        playlist_names.append(playlist_name)
        playlist_ids.append(playlist_id)

playlists =  pd.DataFrame({
    "id" : id_list,
    "playlist_name" : playlist_names,
    "playlist_id" : playlist_ids
})
playlists = playlists[["playlist_name","playlist_id","id"]]
playlists.to_csv("../datos/02 Base de Datos/03_tabla_playlists.csv", index=False)
playlists.sample()

### Crear Tabla

conexion = sql.conectar_bd()
query = ''' CREATE TABLE playlists (
            id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
            playlist_name VARCHAR NOT NULL,
            playlist_id VARCHAR NOT NULL,
            follower_id BIGINT REFERENCES followers(id) ON DELETE CASCADE
            );'''
sql.modificar_bd(conexion,query)

conexion = sql.conectar_bd()
query = '''INSERT INTO playlists(playlist_name,playlist_id,follower_id) VALUES (%s,%s,%s)'''
sql.insertar_muchos_datos(conexion,query,sql.generar_tupla(playlists))

# Tabla Reduced Playlists
- Esta tabla busca tener solo los ids de las playlists que finalmente se usaron para sacar los artistas y géneros.
- De esta forma guardamos todas las playlists en un lado y por otro lado, mostramos las que finalmente usamos.

# Cargar los datasets de las marcas
zara = pd.read_csv('../datos/01 Spotify/01_followers_zara.csv')

# Seleccionar columnas clave de Zara
zara = zara[["brand", "username", "user_id", "playlists"]]

# Convertir las cadenas de texto de la columna "playlists" a diccionarios reales
zara["playlists"] = zara["playlists"].apply(ast.literal_eval)

### Reducir a 10 elementos por diccionario
- Será las playlists que usemos para sacar los géneros

# Lista para almacenar los diccionarios reducidos
reduced_playlists_list = []

# Iterar sobre cada diccionario en la columna "playlists"
for dict_item in zara["playlists"]:
    reduced_dict = {}  # Diccionario reducido
    count = 0  # Contador para elementos procesados

    # Iterar sobre los elementos del diccionario y limitar a 10
    for key, value in dict_item.items():
        if count < 10:
            reduced_dict[key] = value
            count += 1
        else:
            break
    
    # Agregar el diccionario reducido a la lista
    reduced_playlists_list.append(reduced_dict)

# Crear la nueva columna "reduced_playlists" en el DataFrame
zara["reduced_playlists"] = reduced_playlists_list

playlists = zara[["user_id","reduced_playlists"]]
playlists.sample()

### Obtenemos el id de los usuarios haciendo una consulta a la bd

conexion = sql.conectar_bd()
query = '''SELECT * FROM followers'''
df_users = sql.consulta_sql(conexion,query)
df_users = df_users[["id","user_id","username"]]
df_users.sample()

### Realizamos merge y convertimos a diccionario "playlists"

unir = pd.merge(left=df_users,right=playlists,on="user_id")
playlists = unir[["id","reduced_playlists"]]
playlists.sample()

id_list = []
playlist_names = []
playlist_ids = []

for index, fila in playlists.iterrows():
    user_id = fila["id"]
    playlist = fila["reduced_playlists"]

    for playlist_name, playlist_id in playlist.items():
        id_list.append(user_id)
        playlist_names.append(playlist_name)
        playlist_ids.append(playlist_id)

playlists =  pd.DataFrame({
    "id" : id_list,
    "playlist_name" : playlist_names,
    "playlist_id" : playlist_ids
})
playlists = playlists[["playlist_name","playlist_id","id"]]
playlists.to_csv("../datos/02 Base de Datos/04_tabla_reduced_playlists.csv", index=False)
playlists.sample()

### Crear Tabla

conexion = sql.conectar_bd()
query = ''' CREATE TABLE reduced_playlists (
            id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
            playlist_name VARCHAR NOT NULL,
            playlist_id VARCHAR NOT NULL,
            follower_id BIGINT REFERENCES followers(id) ON DELETE CASCADE
        );'''
sql.modificar_bd(conexion,query)

conexion = sql.conectar_bd()
query = '''INSERT INTO reduced_playlists(playlist_name,playlist_id,follower_id) VALUES (%s,%s,%s)'''
sql.insertar_muchos_datos(conexion,query,sql.generar_tupla(playlists))

# Tabla Artistas
- Tiene el nombre del Artista
- Conectado a la playlist de la que viene
- Conectado a la marca del que proviene

artistas = pd.read_csv("../datos/01 Spotify/01 Resumen Marcas/resumen_zara.csv")
artistas

### Nos quedamos con brand y unique_artists

artistas = artistas[["brand","unique_artists"]]
# Convertimos unique_artists a diccionario
artistas["unique_artists"] = artistas["unique_artists"].apply(ast.literal_eval)
artistas

### Crear tabla separando el diccionario

lista_brand = []
nombres_artistas = []
ids_artistas = []

for dictio in artistas["unique_artists"]:
    for artist_id, artist_name in dictio.items():
        lista_brand.append(1) # El id de zara en la bd es 1
        nombres_artistas.append(artist_name)
        ids_artistas.append(artist_id)
artistas = pd.DataFrame({
    "artist_name" : nombres_artistas,
    "artist_id" : ids_artistas,
    "brand_id" : lista_brand
})
artistas.to_csv("../datos/02 Base de Datos/05_tabla_artistas_zara.csv",index=False)
artistas.head()

### Quitar Nulos

artistas.dropna(inplace=True)

### Crear Tabla
- Empezamos a usar text porque es mucho mejor que varchar para nuestro uso

conexion = sql.conectar_bd()
query = ''' create table artists (
            id bigint primary key generated always as identity,
            artist_name text not null,
            artist_id text not null,
            brand_id bigint references brands (id) on delete cascade
            );'''
sql.modificar_bd(conexion,query)

### Insertamos Datos

conexion = sql.conectar_bd()
query = '''INSERT INTO artists(artist_name,artist_id,brand_id) VALUES (%s,%s,%s)'''
sql.insertar_muchos_datos(conexion,query,sql.generar_tupla(artistas))

# Tabla Ranking Artistas
- Tiene el nombre del Artista
- Y el número de veces que aparece un artista
- Conectado a la marca del que proviene

ranking_art = pd.read_csv("../datos/01 Spotify/01 Resumen Marcas/resumen_zara.csv")
ranking_art = ranking_art[["brand","artist_ranking"]]
ranking_art["artist_ranking"] = ranking_art["artist_ranking"].apply(lambda x: ast.literal_eval(x) if pd.notna(x) else [])
ranking_art

lista_brands = []
lista_artistas = []
lista_conteos = []
for tuplas in ranking_art["artist_ranking"]:
    for artista, conteo in tuplas:
        lista_artistas.append(artista)
        lista_conteos.append(conteo)
        lista_brands.append(1) # 1 es zara en la base de datos

ranking_art = pd.DataFrame({
    "artist_name" : lista_artistas,
    "number_of_appearances" : lista_conteos,
    "brand_id" : lista_brands
})

ranking_art.head()

conexion = sql.conectar_bd()
query = ''' create table artists_ranking (
            id bigint primary key generated always as identity,
            artist_name text not null,
            number_of_appearances int not null,
            brand_id bigint references brands (id) on delete cascade
          );'''
sql.modificar_bd(conexion,query)

conexion = sql.conectar_bd()
query = '''INSERT INTO artists_ranking(artist_name,number_of_appearances,brand_id) VALUES (%s,%s,%s)'''
sql.insertar_muchos_datos(conexion,query,sql.generar_tupla(ranking_art))

ranking_art.to_csv("../datos/02 Base de Datos/06_ranking_artistas_zara.csv",index=False)
