==> jupyter-notebooks/08_Como_Usar_Funciones_SQL.ipynb <==
# Como Usar las Funciones
- Aquí podrás ver que espera cada Función!

import sys
sys.path.append("../")
import src.soporte_subida_datos_sql as sp
import pandas as pd
import ast

df = pd.read_csv("../datos/01 Spotify/01 Resumen Marcas/resumen_zara.csv")
df

# obtener_tabla_artistas
- el dataframe el de arriba
- el int, debe ser el brand_id, REVISA BIEN A CUAL MAPEAS!
- Si no existe el brand_id, se te devuelve cuales son los que existen

sp.obtener_tabla_artistas(df,10)

sp.obtener_ranking_artistas(df,10)

# Funcion de 3 Pasos
- Paso uno los mapeamos
- Aquí si quiero ampliar el json con los géneros que no existan

df_step1 = sp.mapear_main_genres(df,10)
df_step1

# Step 2
- Espera el DF que devuelve Step_1 en mapear_main_genres()
- Revisa el brand id, y listo, esta para subir en el paso 3

df_step2 = sp.obtener_main_genres(df_step1,10)
df_step2

# obtener_subgenres
- Espera lo que devuelve mapear_genres()

sp.obtener_subgenres(df_step1,1)
