==> jupyter-notebooks/09_obtener_resto_marcas.ipynb <==
# Extracción del resto de marcas
- Usando las funciones fabricadas a partir de lo aprendido con Zara

import sys
import pandas as pd
sys.path.append("../")
import src.soporte_extraccion_datos as ex

# Primark

primark = pd.read_csv("../datos/01 Spotify/02_followers_primark.csv",index_col=0)

### Extraemos el id de los artistas de las 10 primeras playlists de cada usuario

primark_artists = ex.obtener_id_artistas(primark,output_file="../datos/01 Spotify/00 TempSaves/00_tempsave_primark.csv")
resumen_primark = ex.tabla_resumen(primark_artists,
                                   followers_path="../datos/01 Spotify/02_followers_primark.csv",
                                   resumen_path="../datos/01 Spotify/01 Resumen Marcas/resumen_primark.csv")

resumen_primark = ex.obtener_generos_artistas(resumen_primark,resumen_path="../datos/01 Spotify/01 Resumen Marcas/resumen_primark.csv")
resumen_primark

# Nike

nike = pd.read_csv("../datos/01 Spotify/03_followers_nike.csv",index_col=0)

### Extraemos el id de los artistas de las 10 primeras playlists de cada usuario

nike_artists = ex.obtener_id_artistas(nike,output_file="../datos/01 Spotify/00 TempSaves/00_tempsave_nike.csv")
resumen_nike = ex.tabla_resumen(nike_artists,
                                followers_path="../datos/01 Spotify/03_followers_nike.csv",
                                resumen_path="../datos/01 Spotify/01 Resumen Marcas/resumen_nike.csv")

resumen_nike = ex.obtener_generos_artistas(resumen_nike,resumen_path="../datos/01 Spotify/01 Resumen Marcas/resumen_nike.csv")
resumen_nike

# Adidas

adidas = pd.read_csv("../datos/01 Spotify/04_followers_adidas.csv",index_col=0)

adidas_artists = ex.obtener_id_artistas(adidas,output_file="../datos/01 Spotify/00 TempSaves/00_tempsave_adidas.csv")

adidas_artists

# H&M

hm = pd.read_csv("../datos/01 Spotify/05_followers_hm.csv",index_col=0)

hm_artists = ex.obtener_id_artistas(hm,output_file="../datos/01 Spotify/00 TempSaves/00_tempsave_hm.csv")

hm_artists

# Pull & Bear

pullbear = pd.read_csv("../datos/01 Spotify/06_followers_pull_and_bear.csv",index_col=0)

# Bershka

bershka = pd.read_csv("../datos/01 Spotify/07_followers_bershka.csv",index_col=0)

# Stradivarius

stradivarius = pd.read_csv("../datos/01 Spotify/08_followers_stradivarius.csv",index_col=0)

# Mango

mango = pd.read_csv("../datos/01 Spotify/09_followers_mango.csv",index_col=0)

# Desigual

desigual = pd.read_csv("../datos/01 Spotify/10_followers_desigual.csv",index_col=0)
