==> jupyter-notebooks/10_subir_resto_marcas.ipynb <==
import sys
sys.path.append("../")
import pandas as pd
import src.soporte_subida_datos_sql as sp

# Subir Primark

### Followers

primark = pd.read_csv("../datos/01 Spotify/02_followers_primark.csv")
followers_primark = sp.obtener_tabla_followers(primark,2)
sp.subir_followers(followers_primark)

### Playlists

playlists_primark = sp.obtener_tabla_playlists(primark,2)
sp.subir_playlists(playlists_primark)

### Reduced_playlists

reduced_playlists = sp.obtener_tabla_reduced_playlists(primark,2)
sp.subir_reduced_playlists(reduced_playlists)

resumen_primark = pd.read_csv("../datos/01 Spotify/01 Resumen Marcas/resumen_primark.csv")

### Artistas

artistas = sp.obtener_tabla_artistas(resumen_primark, 2, ruta_csv = "../datos/02 Base de Datos/00_tempsaves/tempsave_artists.csv")
sp.subir_artistas(artistas)

### Ranking Artistas

ranking_artistas = sp.obtener_ranking_artistas(resumen_primark,2)
sp.subir_ranking_artistas(ranking_artistas)

### Main_genres

main_genres = sp.mapear_main_genres(resumen_primark,2)
main_genres = sp.obtener_main_genres(main_genres,2)
sp.subir_main_genres(main_genres)

### Subgenres

subgenres_m = sp.mapear_main_genres(resumen_primark,2)
subgenres = sp.obtener_subgenres(subgenres_m,2)
subgenres.index = subgenres.index + 664
subgenres.reset_index(inplace=True)
sp.subir_subgenres(subgenres,auto_id=True)

# Subir Nike

### Followers

nike = pd.read_csv("../datos/01 Spotify/03_followers_nike.csv")
followers_nike = sp.obtener_tabla_followers(nike,3)
sp.subir_followers(followers_nike)

### Playlists

playlists_nike = sp.obtener_tabla_playlists(nike,3)
sp.subir_playlists(playlists_nike)

### Reduced_playlists

reduced_playlists_nike = sp.obtener_tabla_reduced_playlists(nike,3)
sp.subir_reduced_playlists(reduced_playlists_nike)

resumen_nike = pd.read_csv("../datos/01 Spotify/01 Resumen Marcas/resumen_nike.csv")

### Artistas

artistas = sp.obtener_tabla_artistas(resumen_nike, 3, ruta_csv = "../datos/02 Base de Datos/00_tempsaves/tempsave_artists.csv")
sp.subir_artistas(artistas)

### Ranking Artistas

ranking_artistas = sp.obtener_ranking_artistas(resumen_nike,3)
sp.subir_ranking_artistas(ranking_artistas)

### Main_genres

main_genres = sp.mapear_main_genres(resumen_nike,3)
main_genres = sp.obtener_main_genres(main_genres,3)
sp.subir_main_genres(main_genres)

### Subgenres

subgenres_m = sp.mapear_main_genres(resumen_nike,3)
subgenres = sp.obtener_subgenres(subgenres_m,3)
subgenres.index = subgenres.index + 1324
subgenres.reset_index(inplace=True)
sp.subir_subgenres(subgenres,auto_id=True)
