==> jupyter-notebooks/11_base_datos_usuario.ipynb <==
# Contexto
- La idea es guardar a los usuarios que se unan la primera vez en la base de datos
- Así cuando el usuario vuelva, se comprobara si ya tenemos su información
- Si tenemos su información los datos para el sistema de recomendación se sacaran de supabase
- Si no existe el usuario haremos el codigo de spotipy y lo iremos guardando en la base de datos

# Estos son los comandos utilizados para crear la base de datos

import sys
sys.path.append("../")
import src.soporte_sql as sql
import pandas as pd
import ast

# Tabla Users

conexion = sql.conectar_bd()
query = '''CREATE TABLE users (
            user_id TEXT PRIMARY KEY,
            name TEXT NOT NULL,
            url TEXT UNIQUE NOT NULL,
            email TEXT UNIQUE NOT NULL,
            product_version TEXT NOT NULL
        );'''
sql.modificar_bd(conexion,query)

# Tabla Canciones que le gustan

conexion = sql.conectar_bd()
query = '''CREATE TABLE tracks_user_likes (
            auto_id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
            user_id TEXT REFERENCES users(user_id) ON DELETE CASCADE,
            song_name TEXT NOT NULL,
            song_id TEXT NOT NULL,
            popularity BIGINT NOT NULL,
            song_url TEXT NOT NULL,
            artist_name TEXT NOT NULL,
            artist_id TEXT NOT NULL,
            artist_url TEXT NOT NULL,
            user_added_at TEXT NOT NULL
        );'''
sql.modificar_bd(conexion,query)

# Tabla top_tracks

conexion = sql.conectar_bd()
query = '''CREATE TABLE top_tracks (
            auto_id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
            user_id TEXT REFERENCES users(user_id) ON DELETE CASCADE,
            ranking INT NOT NULL,
            song_name TEXT NOT NULL,
            song_id TEXT NOT NULL,
            popularity BIGINT NOT NULL,
            song_url TEXT NOT NULL,
            artist_name TEXT NOT NULL,
            artist_id TEXT NOT NULL,
            artist_url TEXT NOT NULL
        );'''
sql.modificar_bd(conexion,query)

# Tabla user_artists_ranking

conexion = sql.conectar_bd()
query = '''CREATE TABLE user_artists_ranking (
            auto_id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
            user_id TEXT REFERENCES users(user_id) ON DELETE CASCADE,
            artist_name TEXT NOT NULL,
            number_of_appearances INT NOT NULL
        );'''
sql.modificar_bd(conexion,query)

# Tabla user_main_genres

conexion = sql.conectar_bd()
query = '''CREATE TABLE user_main_genres (
            auto_id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
            user_id TEXT REFERENCES users(user_id) ON DELETE CASCADE,
            genre_name TEXT NOT NULL,
            number_of_appearances INT NOT NULL
        );'''
sql.modificar_bd(conexion,query)

# Tabla user_subgenres

conexion = sql.conectar_bd()
query = '''CREATE TABLE user_subgenres (
            auto_id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
            user_id TEXT REFERENCES users(user_id) ON DELETE CASCADE,
            subgenre_name TEXT NOT NULL,
            number_of_appearances INT NOT NULL
        );'''
sql.modificar_bd(conexion,query)
