==> jupyter-notebooks/12_obtener_datos_ejemplo.ipynb <==
# Contexto
- Utilizando la playlist de "Canciones que te gustan" obtenemos todos los tracks de ahí
- También obtendremos todos los "Top tracks" por si el usuario no tuviera de lo otro
- La idea es sacar lo mismo que en las marcas:
    - Top artistas
    - Top géneros
    - Top subgéneros
### Esto es un ejemplo de las funciones que se llaman en streamlit

import sys
sys.path.append("../")
import pandas as pd
import src.soporte_spotify as api
import src.soporte_sql as sql
import src.soporte_spotify_streamlit as spot
import os
from supabase import create_client, Client
from dotenv import load_dotenv
# Cargar variables de entorno
load_dotenv()

sp = api.load_credentials()

# Configuración de Supabase
url = os.getenv("project_url")
key = os.getenv("browser_safe_key")
supabase: Client = create_client(url, key)

# Añadir y subir datos de usuario

spot.get_current_user(sp)

# Obtener todas las canciones que están en "Canciones que te gustan"

spot.get_all_saved_tracks(sp)

# Obtener canciones más escuchadas

spot.get_all_top_tracks(sp)

# Sacar ranking artistas

spot.get_user_artist_ranking(sp, supabase)

# Obtener ranking géneros y subgeneros

spot.get_user_genre_and_subgenre_ranking(sp,supabase)
